renderer separated
