+++ experimental/pyglet_test/graphics_experiment.py
